refactor(dto): CursorData, CursorResponse Dto를 common/dto 패키지에 CursorDto를 생성하여 이동
- 추후에 마일스톤이나 레이블에서도 무한스크롤이 적용될 수도 있을 것 같아 이동
