+++ backend/src/main/java/codesquad/team01/issuetracker/issue/dto/IssueDto.java
@@ -6,22 +6,19 @@ import java.util.ArrayList;
 import java.util.Base64;
 import java.util.List;
 
-import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.databind.SerializationFeature;
 import com.fasterxml.jackson.datatype.jsr310.JavaTimeModule;
 
-import codesquad.team01.issuetracker.common.exception.CursorException;
+import codesquad.team01.issuetracker.common.dto.CursorDto;
 import codesquad.team01.issuetracker.issue.domain.IssueState;
 import codesquad.team01.issuetracker.label.dto.LabelDto;
 import codesquad.team01.issuetracker.milestone.dto.MilestoneDto;
 import codesquad.team01.issuetracker.user.dto.UserDto;
 import jakarta.validation.constraints.Pattern;
 import jakarta.validation.constraints.Positive;
-import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
-import lombok.NoArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 
 @Slf4j
@@ -96,7 +93,7 @@ public class IssueDto {
 			return state != null ? state : "open";
 		}
 
-		public CursorData decode() {
+		public CursorDto.CursorData decode() {
 			if (cursor == null || cursor.isBlank()) {
 				return null;
 			}
@@ -106,7 +103,7 @@ public class IssueDto {
 				ObjectMapper mapper = new ObjectMapper()
 					.registerModule(new JavaTimeModule())
 					.disable(SerializationFeature.WRITE_DATES_AS_TIMESTAMPS);
-				return mapper.readValue(decoded, CursorData.class);
+				return mapper.readValue(decoded, CursorDto.CursorData.class);
 			} catch (Exception e) {
 				log.warn("커서 디코딩 실패", e);
 				// 디코딩 실패 시 null 반환 -> 첫 페이지
@@ -123,7 +120,7 @@ public class IssueDto {
 	public record ListResponse(
 		int totalCount,
 		List<ListItemResponse> issues,
-		CursorResponse cursor
+		CursorDto.CursorResponse cursor
 	) {
 	}
 
@@ -145,13 +142,6 @@ public class IssueDto {
 
 		@Builder.Default
 		private final List<LabelDto.FilterListItemResponse> labels = new ArrayList<>();
-	}
-
-	@Builder
-	public record CursorResponse(
-		String next,
-		boolean hasNext
-	) {
 	}
 
 	// 상태별 이슈 개수 응답 DTO
@@ -239,27 +229,4 @@ public class IssueDto {
 		}
 	}
 
-	@Getter
-	@Builder
-	@AllArgsConstructor
-	@NoArgsConstructor
-	public static class CursorData {
-		private Integer id;
-		private LocalDateTime createdAt;
-
-		public String encode() {
-			try {
-				String json = new ObjectMapper()
-					.registerModule(new JavaTimeModule())
-					.disable(SerializationFeature.WRITE_DATES_AS_TIMESTAMPS)
-					.writeValueAsString(this);
-
-				return Base64.getEncoder().encodeToString(json.getBytes(StandardCharsets.UTF_8));
-			} catch (JsonProcessingException e) { // 인코딩 실패 시
-				throw new CursorException("json 직렬화 실패");
-			} catch (Exception e) {
-				throw new CursorException(e.getMessage());
-			}
-		}
-	}
 }
