+++ backend/src/main/java/codesquad/team01/issuetracker/issue/repository/IssueQueryRepository.java
@@ -2,6 +2,7 @@ package codesquad.team01.issuetracker.issue.repository;
 
 import java.util.List;
 
+import codesquad.team01.issuetracker.common.dto.CursorDto;
 import codesquad.team01.issuetracker.issue.domain.IssueState;
 import codesquad.team01.issuetracker.issue.dto.IssueDto;
 
@@ -11,7 +12,7 @@ public interface IssueQueryRepository {
 
 	List<IssueDto.BaseRow> findIssuesWithFilters(
 		IssueState state, Integer writerId, Integer milestoneId,
-		List<Integer> labelIds, List<Integer> assigneeIds, IssueDto.CursorData cursor);
+		List<Integer> labelIds, List<Integer> assigneeIds, CursorDto.CursorData cursor);
 
 	IssueDto.CountResponse countIssuesWithFilters(Integer writerId, Integer milestoneId,
 		List<Integer> labelIds, List<Integer> assigneeIds);
